+++ lawfastapi/app/rag_chain.py
@@ -1,5 +1,5 @@
-from rag.pdf import PDFRetrievalChain
-from rag.utils import format_docs, format_searched_docs
+from retrieval_chain.pdf import PDFRetrievalChain
+from retrieval_chain.utils import format_docs, format_searched_docs
 from schema.graph_state import GraphState
 from langchain_upstage import UpstageGroundednessCheck
 from langchain_core.prompts import ChatPromptTemplate
